[Testing] Reverting move of pilot analysis_test.go
Fixes #27951
